# openapi.yaml
openapi: 3.1.0
info:
  title: API SnipMe
  description: |
    API REST para acortador de URLs con autenticación OAuth2 y JWT.

    ## Autenticación

    La API utiliza OAuth2 con Google y GitHub para autenticación, y JWT Bearer tokens para autorización.

    ### Flujo de autenticación:
    1. Redirigir al usuario a `/auth/google/login` o `/auth/github/login`
    2. El usuario autoriza la aplicación en el proveedor OAuth
    3. La API retorna un `access_token` (JWT) y `refresh_token`
    4. Usar el `access_token` en el header `Authorization: Bearer <token>` para endpoints protegidos
    5. Renovar el token con `/auth/refresh` cuando expire

  version: 0.1.0
  contact:
    name: Alex Perez Garcia
    email: aperezgdev@gmail.com
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Servidor local de desarrollo
  - url: https://api.snipme.app
    description: Servidor de producción
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtenido mediante OAuth2
  schemas:
    AuthenticationResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT token para autenticación
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          format: uuid
          description: Token para renovar el access_token
          example: 550e8400-e29b-41d4-a716-446655440000
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: Tiempo de expiración en segundos
          example: 900
    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          description: Refresh token obtenido en el login
          example: 550e8400-e29b-41d4-a716-446655440000
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Invalid request
        message:
          type: string
          example: The request body is invalid
paths:
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Renovar access token
      description: Renueva el access_token usando un refresh_token válido
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token renovado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'
        '400':
          description: Refresh token inválido o expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...